size_t is definitivelly a strange type. There is no way to print it correctly on all platforms. So here is the trick: I decide to print it always as an unsigned long and explicitly cast everything to this type. Thus, I change all printf formats from %d to %lu and cast all arguemnts to the correct type (unsigned long).
This commit was SVN r5568.

+++ src/tools/orterun/orterun.c
@@ -398,8 +398,8 @@ static void dump_aborted_procs(orte_jobid_t jobid)
                 ++num_killed;
             } else {
                 if (num_aborted < max_display_aborted) {
-                    fprintf(stderr, "Job rank %ld (pid %ld) on node \"%s\" exited on signal %d\n",
-                            rank, pid, node_name, WTERMSIG(exit_status));
+                    fprintf(stderr, "Job rank %lu (pid %lu) on node \"%s\" exited on signal %d\n",
+                            (unsigned long)rank, (unsigned long)pid, node_name, WTERMSIG(exit_status));
                 }
                 ++num_aborted;
             }
